# mat: optimize multri