optimize the paths for diagonal matrices in MulTri.  there was a TODO and i use tri*diag a lot.
add benchmarks for multri

the performance improvement is pretty ridiculous as it cuts 1 or 2 dimensions out of the problem.

<table class='benchstat oldnew'>
<tr class='configs'><th><th>old.txt<th>new.txt
<tbody>
<tr><th><th colspan='2' class='metric'>time/op<th>delta
<tr class='unchanged'><td>TriMul1000-4<td>2.77s ± 6%<td>2.79s ± 3%<td class='nodelta'>~<td class='note'>(p=0.549 n=10&#43;9)
<tr class='better'><td>TriMulDiag1000-4<td>1.71s ± 6%<td>0.01s ± 3%<td class='delta'>−99.53%<td class='note'>(p=0.000 n=9&#43;9)
<tr class='better'><td>TriMul2Diag1000-4<td>1.30s ± 2%<td>0.00s ±15%<td class='delta'>−99.96%<td class='note'>(p=0.000 n=9&#43;10)
<tr><td>&nbsp;
</tbody>
</table>

<!--
Checklist:

- API changes have been discussed
- code is goformated correctly (goimports)
- packages with generated code have had code generation run
- tests pass locally
- linked to relevant issues

Please make sure your commit message summary line and pull request
title match the Go convention; a one-line summary of the change,
prefixed by the primary affected package that should complete the
sentence, "This change modifies Gonum to _____."
-->