@jonreiter 2019-04-16
@sbinet 
stylistic changes fair enough and made.
re benchmark i added that addition to the description. is that what you meant?