@kortschak 2019-04-16
Nice speedups.