@jonreiter 2019-04-16
@kortschak 
the non-diagonal timings are a bit noisy to the point my laptop is hot.  the diag timings are however quite stable.  type switches amended too.

<table class='benchstat oldnew'>
<tr class='configs'><th><th>old.txt<th>new.txt
<tbody>
<tr><th><th colspan='2' class='metric'>time/op<th>delta
<tr class='unchanged'><td>TriMul/BenchmarkMulTri100-4<td>2.16ms ± 2%<td>2.20ms ± 5%<td class='nodelta'>~<td class='note'>(p=0.094 n=9&#43;9)
<tr class='unchanged'><td>TriMul/BenchmarkMulTri200-4<td>16.9ms ± 2%<td>17.0ms ± 1%<td class='nodelta'>~<td class='note'>(p=0.243 n=10&#43;9)
<tr class='worse'><td>TriMul/BenchmarkMulTri400-4<td>139ms ± 2%<td>142ms ± 3%<td class='delta'>&#43;2.41%<td class='note'>(p=0.009 n=10&#43;10)
<tr class='unchanged'><td>TriMul/BenchmarkMulTri800-4<td>1.29s ± 4%<td>1.32s ± 6%<td class='nodelta'>~<td class='note'>(p=0.356 n=9&#43;10)
<tr class='worse'><td>TriMul/BenchmarkMulTri1600-4<td>12.8s ± 3%<td>13.8s ± 4%<td class='delta'>&#43;7.27%<td class='note'>(p=0.000 n=9&#43;10)
<tr class='better'><td>TriMulDiag/BenchmarkMulTri100-4<td>1.78ms ± 1%<td>0.08ms ± 3%<td class='delta'>−95.56%<td class='note'>(p=0.000 n=9&#43;10)
<tr class='better'><td>TriMulDiag/BenchmarkMulTri200-4<td>13.8ms ± 3%<td>0.3ms ± 2%<td class='delta'>−97.66%<td class='note'>(p=0.000 n=9&#43;9)
<tr class='better'><td>TriMulDiag/BenchmarkMulTri400-4<td>110ms ± 4%<td>1ms ± 1%<td class='delta'>−98.86%<td class='note'>(p=0.000 n=10&#43;9)
<tr class='better'><td>TriMulDiag/BenchmarkMulTri800-4<td>857ms ± 1%<td>5ms ± 1%<td class='delta'>−99.40%<td class='note'>(p=0.000 n=9&#43;10)
<tr class='better'><td>TriMulDiag/BenchmarkMulTri1600-4<td>7.02s ± 6%<td>0.02s ± 3%<td class='delta'>−99.71%<td class='note'>(p=0.000 n=10&#43;10)
<tr class='better'><td>TriMul2Diag/BenchmarkMulTri100-4<td>1.43ms ± 3%<td>0.00ms ± 2%<td class='delta'>−99.75%<td class='note'>(p=0.000 n=10&#43;9)
<tr class='better'><td>TriMul2Diag/BenchmarkMulTri200-4<td>11.0ms ± 1%<td>0.0ms ± 1%<td class='delta'>−99.91%<td class='note'>(p=0.000 n=8&#43;10)
<tr class='better'><td>TriMul2Diag/BenchmarkMulTri400-4<td>84.7ms ± 2%<td>0.0ms ± 4%<td class='delta'>−99.94%<td class='note'>(p=0.000 n=9&#43;10)
<tr class='better'><td>TriMul2Diag/BenchmarkMulTri800-4<td>670ms ± 2%<td>0ms ± 2%<td class='delta'>−99.97%<td class='note'>(p=0.000 n=9&#43;9)
<tr class='better'><td>TriMul2Diag/BenchmarkMulTri1600-4<td>5.33s ± 2%<td>0.00s ± 1%<td class='delta'>−99.97%<td class='note'>(p=0.000 n=9&#43;10)
<tr><td>&nbsp;
</tbody>

</table>